split cache_streaming related into cache_streaming module
